\name{concave}
\docType{data}
\alias{concave}
\alias{concave.cid}
\title{The concave dataset}
\description{A simulated flowcytometry data with two concave shapes}
\usage{data(concave)}
\format{
  An object (concave) of data frame with  rows and 3 columns and a
  vector (concave.cid) for the true cluster labels. 
}
\references{
  Ge Y. et al, flowPeaks: a fast unsupervised clustering for flow
  cytometry data via K-means and density peak finding, 2012, Bioinformatics, in press.
  }
\keyword{datasets}
