\name{print.flowPeaks}
\alias{print.flowPeaks}
\title{The display of the flowPeaks results}
\description{The display of the flowPeaks results}
\usage{
\method{print}{flowPeaks}(x,\dots)
}
\arguments{
  \item{x}{The output from the function \code{\link{flowPeaks}}}
  \item{\dots}{Optional additional arguments.  At present no additional
    arguments are used.}
}
\author{
  Yongchao Ge \email{yongchao.ge@gmail.com}
}
\seealso{\code{\link{flowPeaks}}} 
\keyword{print}
