\name{summary.flowPeaks}
\alias{summary.flowPeaks}
\title{The summary of the flowPeaks results}
\description{The summary of the flowPeaks results}
\usage{
\method{summary}{flowPeaks}(object,\dots)
}
\arguments{
  \item{object}{The output from the function \code{\link{flowPeaks}}}
  \item{\dots}{Optional additional arguments.  At present no additional
    arguments are used.}
}
\author{
  Yongchao Ge \email{yongchao.ge@gmail.com}
}
\seealso{\code{\link{flowPeaks}}} 
\keyword{summary}
